tpch: support int64 orderkey
With scale factors greater than 300, the l_orderkey column in the lineitem
table exceeds the capacity of a 32-bit integer. This patch increases it to
64-bit.

I also changed it to a signed integer, because we plan to eventually remove
unsigned ints. Thus, it makes more sense to benchmark the signed variant.

I also added better error messages to CHECK failures, since the overflow wasn't
obvious to diagnose.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6438
Tested-by: jenkins

+++ src/kudu/benchmarks/tpch/rpc_line_item_dao-test.cc
@@ -56,7 +56,7 @@ class RpcLineItemDAOTest : public KuduTest {
   KuduSchema schema_;
 
   static void BuildTestRow(int order, int line, KuduPartialRow* row) {
-    CHECK_OK(row->SetUInt32(tpch::kOrderKeyColIdx, order));
+    CHECK_OK(row->SetInt64(tpch::kOrderKeyColIdx, order));
     CHECK_OK(row->SetUInt32(tpch::kLineNumberColIdx, line));
     CHECK_OK(row->SetUInt32(tpch::kPartKeyColIdx, 12345));
     CHECK_OK(row->SetUInt32(tpch::kSuppKeyColIdx, 12345));
@@ -75,7 +75,7 @@ class RpcLineItemDAOTest : public KuduTest {
   }
 
   static void UpdateTestRow(int key, int line_number, int quantity, KuduPartialRow* row) {
-    CHECK_OK(row->SetUInt32(tpch::kOrderKeyColIdx, key));
+    CHECK_OK(row->SetInt64(tpch::kOrderKeyColIdx, key));
     CHECK_OK(row->SetUInt32(tpch::kLineNumberColIdx, line_number));
     CHECK_OK(row->SetUInt32(tpch::kQuantityColIdx, quantity));
   }
